Palete Tiles and player interaction
Creation of the game's initial map through the Tiles Palette, available from the 2017 version of Unity. Animations were also created with the assets, for example, torches. Finally, some character fixes and addition of interaction to objects. There was also a change in the physics of the player, now there is no more friction with the scenario.