Player weapons and first enemy
Adding a weapon to the player and creating a static enemy for testing purposes. A list of damage types has also been added to the GameController. In the future, depending on the enemy, when the weapon hits, it may have a damage bonus if the weapon has a special ability and the enemy is fulnerable to it and vice versa