Scenario part 2
Added part 2 of the scenario. Character starts on the island and with green vegetation and after passing, there is the second part, where the jumps on the skewers begin and after passing, the bridge of wood and bricks. At the end of the bridge, there is the entrance to the Mine through the main stairs.